xilpm: versal: server: Fix MISRA 12.2 violation
Fix below MISRA-C violation:

- misra_c_2012_rule_12_2_violation:
	The right hand operand of a shift operator shall lie
	in the range zero to one less than the width in bits


Acked-for-series: Jyotheeswar Reddy Mutthareddyvari <jyotheeswar.reddy.mutthareddyvari@xilinx.com>

+++ lib/sw_services/xilpm/src/versal/server/xpm_clock.c
@@ -788,8 +788,8 @@ XStatus XPmClock_QueryTopology(u32 ClockId, u32 Index, u32 *Resp)
 			goto done;
 		}
 		Resp[0] = (u32)TYPE_PLL;
-		Resp[0] |=	CLK_SET_RATE_NO_REPARENT << CLK_CLKFLAGS_SHIFT;
-		Resp[0] |=	NA_TYPE_FLAGS <<	CLK_TYPEFLAGS_SHIFT;
+		Resp[0] |= CLK_SET_RATE_NO_REPARENT << CLK_CLKFLAGS_SHIFT;
+		Resp[0] |= ((u32)NA_TYPE_FLAGS) << CLK_TYPEFLAGS_SHIFT;
 	} else {
 		Status = XST_FAILURE;
 		goto done;
